Ebird Scraping Related/requirements.txt
google-cloud-storage==2.6.0
python-dotenv==0.21.0
pandas==1.4.2